Added PosAdapters

+++ MicroService/invoice-service/src/main/java/com/kodlamaio/invoiceservice/business/responses/CreateInvoiceResponse.java
@@ -17,7 +17,7 @@ public class CreateInvoiceResponse {
     private Date creationDate;
     private LocalDate startDate;
     private LocalDate finishDate;
-    private short totalRentalDay;
+    private int totalRentalDay;
     private double priceOfDays;
     private double rentalCarTotalPrice;
     private int rentalCarId;
